+++ src/systemEntities/email.js
@@ -1,0 +1,8 @@
+/**
+ * @author wingbot.ai
+ */
+'use strict';
+
+module.exports = ['email', /(?<=(\s|^|:))[a-zA-Z0-9!#$%&'*+\-=?^_`{|}~"][^@:\s]*@[^.@\s]+\.[^@\s,]+/, {
+    anonymize: true
+}];