+++ src/systemEntities/index.js
@@ -1,0 +1,12 @@
+/**
+ * @author David Menger
+ */
+'use strict';
+
+const email = require('./email');
+const phone = require('./phone');
+
+module.exports = [
+    email,
+    phone
+];