+++ src/wingbot/WingbotModel.js
@@ -3,6 +3,7 @@
 const { default: fetch } = require('node-fetch');
 const assert = require('assert');
 const CachedModel = require('./CachedModel');
+const systemEntities = require('../systemEntities');
 
 const DEFAULT_MATCHES = 3;
 const SERVICE_URL = 'https://model.wingbot.ai';
@@ -56,6 +57,11 @@ class WingbotModel extends CachedModel {
 
         this._serviceUrl = options.serviceUrl || SERVICE_URL;
         this._model = options.model;
+
+        // apply default entities
+        systemEntities
+            // @ts-ignore
+            .forEach(([name, d, opts = {}]) => this.setEntityDetector(name, d, opts));
     }
 
     /**