+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.7.0] - 2020-11-16
+
+ional entity in quick reply accepts also entities in state
+
+### Added
+
+- system entities for `email` and `phone`
+
 ## [3.6.1] - 2020-11-10
 
 ### Changed